map: We used map to modify the stock price value synchronously. In the getPriceStream method, we transform the price by adding 2 to it. Similarly, in the getSinglePrice method, we added 5 to the price.
filter: Applied the filter operator to only select stock prices greater than 148 in both the getPriceStream and fetchAdditionalStockData methods.

flatMap: In the fetchAdditionalStockData method, we simulate an asynchronous operation (like an API call) using flatMap. We delay the price transformation by 500ms to simulate an API call, and then modify the price. flatMap is useful here because we are dealing with asynchronous streams.